backend/finapp/requirements.txt
Django
djangorestframework
mysqlclient
djangorestframework-simplejwt
django-cors-headers
